# docs/openapi.yaml
openapi: 3.0.0
info:
  description: UdaConnect Project Cloud Native Arch
  version: "1.0.0"
  title: Udaconnect API
  contact:
    email: devi.25@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
paths:
...
  /persons/{personId}/connection:
    get:
      summary: Connection details for person
      operationId: personConnections
      parameters:
        - in: path
          name: personId
          description: ID of a person to show
          schema:
            type: integer
            format: int32
            minimum: 0
          required: true
        - in: query
          name: start_date
          description: number of records to skip for pagination
          schema:
            type: string
            format: date-time
          required: true
        - in: query
          name: end_date
          description: number of records to skip for pagination
          schema:
            type: string
            format: date-time
          required: true
        - in: query
          name: distance
          description: distance in meters to permit for connection search
          schema:
            type: integer
            format: int32
            minimum: 0
          required: true
        - in: query
          name: skip
          description: number of records to skip for pagination
          schema:
            type: integer
            format: int32
            minimum: 0
        - in: query
          name: limit
          description: maximum number of records to return
          schema:
            type: integer
            format: int32
            minimum: 0
            maximum: 50
      responses:
        '200':
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Connection'
        '400':
          description: Invaid Input Parameter
...
components:
  schemas:
    Person:
      type: object
      required:
        - id
        - firstName
        - lastName
        - companyName
      properties:
        id:
          type: integer
          format: number
          example: 1
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Peter"
        companyName: 
          type: string
          example: "Google Co."
    Location:
      type: object
      required:
        - id
        - person
        - coordinate
        - creationTime
      properties:
        id:
          type: integer
          example: 1
        person:
          $ref: '#/components/schemas/Person'
        creationTime:
          type: string
          format: date-time
          example: '2016-08-29T09:12:33.001Z'
        latitude:
          type: number
          example: 41.40338
        longitude:
          type: number
          example: 2.17403
    Connection:
      type: object
      required:
        - person
        - location
      properties:
        person:
          $ref: '#/components/schemas/Person'
        location:
          $ref: '#/components/schemas/Location'